requirements.txt
torch
torchvision
numpy==1.25
tqdm
tensorboard
matplotlib
pandas

speechbrain
datasets
torch_audiomentations
editdistance
wandb
pyctcdecode
torchaudio
pillow
pytest
python-dotenv
rotary-embedding-torch
https://github.com/kpu/kenlm/archive/master.zip
torchmetrics
pyloudnorm
fire
transformers
accelerate
pesq
soundfile
safetensors
Unidecode
gdown
lovely-tensors
hydra-core
inflect
pyworld